api/requirements.txt
sanic==19.3.1
sanic-cors==0.9.8
python-box==3.4.1
jinja2==2.10.1
web3==4.9.2